feat(server): change NewInstance to return error on version change
Previously, the `NewInstance` method in `db-instance-manager` would silently change
the version of a PostgreSQL instance if a different version was provided. This
behavior could lead to unexpected results.

Now, `NewInstance` will return an error if an attempt is made to change the
version of an existing PostgreSQL instance. This change makes the behavior of
`NewInstance` more predictable and helps prevent potential issues from
silently changing the version of a PostgreSQL instance.