refactor(server): replace 'handler' package with 'server'
The 'handler' package was renamed to 'server' to better reflect its purpose. All
references to 'handler' in the codebase have been updated to 'server'. This
includes changes in function signatures, type definitions, and import statements.
The 'handler' package has been deleted and a new 'server' package has been
created. This change does not affect the functionality of the code.