feat(e2e): Add support for different compose providers
This commit introduces the ability to run end-to-end tests using
different compose providers such as 'docker-compose' and
'podman-compose'. The changes include updates to the test files and
the just scripts to accommodate the new feature. This will allow
developers to run tests in different environments based on their
preferred compose provider.