feat(e2e): Refactor secrets management for e2e tests
This commit refactors the secrets management for end-to-end tests.
The secrets are now named with a consistent prefix and are created
as external secrets in the docker-compose files.

This fix the secrets can't be read by user other than root issue.
Note that swarm mode is required to use external secrets when using
docker-compose.

And we use podman-compose in the e2e tests by default.