feat(e2e): create secret support both docker-compose and podman-compose
This commit updates the uid and gid values in the pg-helper.yaml file to
be strings instead of integers. This is to ensure compatibility with
systems that expect these values to be strings.

Additionally, the method for checking if a secret exists in the
secrets.just file has been improved. Instead of directly checking if
the secret exists, the code now lists all secrets and checks if the
desired secret is in that list. This provides a more robust check.