feat(README): add details about compose providers
This commit adds a new section in the README about the three compose
providers: `docker-compose`, `podman-compose`, and `nerdctl compose`.
It includes a comparison table and explains why `podman-compose` is
chosen as the default provider due to its support for external secrets
with specific user, group, and mode. Further investigation is needed
for `docker-compose` and `nerdctl compose` does not support these
features yet.